Revert "refactor `curve` adaptor"
