+++ src/adaptors/curve/index.js
@@ -1,274 +1,269 @@
-const { default: BigNumber } = require('bignumber.js');
-
 const utils = require('../utils');
+const pools = require('./pools');
 
-const legacy = require('./legacy-adaptor');
-const {
-  CRV_API_BASE_URL,
-  BLOCKCHAINIDS,
-  BLOCKCHAINID_TO_REGISTRIES,
-} = require('./config');
+// comes from different locations, merging func
+const mergeBaseApy = async () => {
+  // this is the base apy data, split up into 2 different json files
+  let dataBaseApy = await utils.getData(
+    'https://stats.curve.fi/raw-stats/apys.json'
+  );
+  dataBaseApy = dataBaseApy.apy.day;
 
-const getPools = async (blockchainId) => {
-  const poolsByAddress = {};
-  for (const registry of BLOCKCHAINID_TO_REGISTRIES[blockchainId]) {
-    const uri = `/getPools/${blockchainId}/${registry}`;
-    let response;
-    try {
-      response = await utils.getData(CRV_API_BASE_URL + uri);
-    } catch (error) {
+  let dataBaseApyOther = await utils.getData(
+    'https://stats.curve.fi/raw-stats-crypto/apys.json'
+  );
+  dataBaseApyOther = dataBaseApyOther.apy.week;
+
+  const keys = Object.keys(dataBaseApyOther);
+
+  for (const i of keys) {
+    dataBaseApy[i] = dataBaseApyOther[i];
+  }
+
+  return [dataBaseApy, keys];
+};
+
+const getDataEth = async () => {
+  const poolStats = [];
+
+  // 1) get price data
+  const pricesUSD = await utils.getCGpriceData(
+    'ethereum,bitcoin,chainlink,stasis-eurs',
+    true
+  );
+
+  // 2) get complete base apy data (this does not include any of the factory pool data...)
+  const [dataBaseApy, keysExtra] = await mergeBaseApy();
+
+  // 3) pull curve pool stats data and prepare the individual pool data
+  for (const [pool, apy] of Object.entries(dataBaseApy)) {
+    // there are few pools in the data which are actually not
+    // displayed on the frontend, I remove them
+    // here, probably not the best idea hardcoding this...
+    if ('rens,linkusd,idle'.includes(pool)) {
       continue;
     }
-    if (response?.success) {
-      const poolsByAddressForRegistry = Object.fromEntries(
-        response.data.poolData.map((pool) => [pool.address, pool])
-      );
-      Object.assign(poolsByAddress, poolsByAddressForRegistry);
-    }
-  }
-  return poolsByAddress;
-};
 
-const getSubGraphData = async (blockchainId) => {
-  const uri = `/getSubgraphData/${blockchainId}`;
-  let response;
-  try {
-    response = await utils.getData(CRV_API_BASE_URL + uri);
-  } catch (error) {
-    return {};
-  }
-  if (response?.success) {
-    const poolSubgraphsByAddress = Object.fromEntries(
-      response.data.poolList.map((pool) => [pool.address, pool])
-    );
-    return poolSubgraphsByAddress;
-  } else {
-    return {};
-  }
-};
-
-const getGaugesByChain = async () => {
-  const gaugeUri = '/getGauges';
-  let gaugeResponse;
-  try {
-    gaugeResponse = await utils.getData(CRV_API_BASE_URL + gaugeUri);
-  } catch (error) {
-    return {};
-  }
-
-  const blockchainRegExps = BLOCKCHAINIDS.filter(
-    (blockchainId) => blockchainId !== 'ethereum'
-  ).map((blockchainId) => new RegExp(`^(${blockchainId})-(.+)`));
-
-  const gaugesDataByChain = Object.fromEntries(
-    BLOCKCHAINIDS.map((blockchainId) => [blockchainId, {}])
-  );
-  for (const [gaugeName, gaugeDatum] of Object.entries(
-    gaugeResponse.data.gauges
-  )) {
-    let match = null;
-    for (const re of blockchainRegExps) {
-      match = gaugeName.match(re);
-      if (match) {
-        break;
-      }
-    }
-    // if no match, it's an ethereum gauge
-    let blockchainId;
-    let _gaugeName;
-    if (!match) {
-      _gaugeName = gaugeName;
-      blockchainId = 'ethereum';
+    if (keysExtra.includes(pool)) {
+      url = `https://stats.curve.fi/raw-stats-crypto/${pool}-1440m.json`;
     } else {
-      blockchainId = match[1];
-      _gaugeName = match[2];
+      url = `https://stats.curve.fi/raw-stats/${pool}-1440m.json`;
     }
-    const _gaugeDatum = Object.assign(
-      {
-        gaugeName: _gaugeName,
-      },
-      gaugeDatum
+    let poolData = await utils.getData(url);
+    const maxTimestamp = Math.max(...poolData.map((el) => el.timestamp), 0);
+    poolData = poolData.find((el) => el.timestamp === maxTimestamp);
+
+    // the price will be different depending on pool type
+    // usd pools are 1, btc pools need to be multiplied by the btc price etc
+    // cause total supply is in the native unit, like balance
+    if (pool.includes('eth')) {
+      price = pricesUSD.ethereum.usd;
+    } else if (pool.includes('btc')) {
+      price = pricesUSD.bitcoin.usd;
+    } else if (pool.includes('link')) {
+      price = pricesUSD.chainlink.usd;
+    } else if (pool.includes('eur')) {
+      price = pricesUSD['stasis-eurs'].usd;
+    } else {
+      price = 1;
+    }
+
+    // for tricrypto pools i use the balances and prices cause
+    // supply cant just be multiplied by a single price
+    const scalingFactor = 1e18;
+    poolData.virtual_price /= scalingFactor;
+    poolData.supply /= scalingFactor;
+
+    if (keysExtra.includes(pool)) {
+      if (pool.includes('tricrypto')) {
+        decimals = [1e6, 1e8, 1e18]; // usdt, btc, eth
+      } else if (pool.includes('eurtusd')) {
+        decimals = [1e6, 1e18];
+      } else if (pool.includes('eursusd')) {
+        decimals = [1e6, 1e2];
+      } else if (pool.includes('crveth')) {
+        decimals = [1e18, 1e18];
+      } else if (pool.includes('cvxeth')) {
+        decimals = [1e18, 1e18];
+      } else if (pool.includes('xautusd')) {
+        decimals = [1e6, 1e18];
+      } else if (pool.includes('spelleth')) {
+        decimals = [1e18, 1e18];
+      } else if (pool.includes('teth')) {
+        decimals = [1e18, 1e18];
+      }
+
+      if (pool === 'eurtusd') {
+        tvl = (poolData.balances[0] / decimals[0]) * poolData.price_scale;
+        tvl += poolData.balances[1] / decimals[1];
+      } else if (pool === 'eursusd') {
+        tvl = (poolData.balances[0] / decimals[0]) * poolData.price_scale;
+        tvl += poolData.balances[1] / decimals[1];
+      } else if (pool === 'xautusd') {
+        tvl = (poolData.balances[0] / decimals[0]) * poolData.price_scale;
+        tvl += poolData.balances[1] / decimals[1];
+      } else {
+        tvl = 0;
+        poolData.balances.forEach((el, i) => {
+          tvl += (el / decimals[i]) * poolData.crypto_prices[i];
+        });
+      }
+    } else {
+      // for all else
+      tvl = poolData.supply * price;
+    }
+
+    // scale by virtual price
+    tvl *= poolData.virtual_price;
+
+    let apyBase = apy * 100;
+    // some of the values are negative, set to 0
+    apyBase = apyBase < 0 ? 0 : apyBase;
+
+    // 4) get crv+reward apr data
+    let dataCrvApy = await utils.getData(
+      'https://www.convexfinance.com/api/curve-apys'
     );
-    // gauge address on pools crv API is lowercase
-    gaugesDataByChain[blockchainId][
-      _gaugeDatum.gauge.toLowerCase()
-    ] = _gaugeDatum;
-  }
+    let dataRewardApy = await utils.getData('https://api.curve.fi/api/getApys');
+    dataCrvApy = dataCrvApy.apys;
+    dataRewardApy = dataRewardApy.data;
 
-  return gaugesDataByChain;
-};
-
-const getMainPoolGaugeRewards = async () => {
-  const uri = '/getMainPoolsGaugeRewards';
-  let response;
-  try {
-    response = await utils.getData(CRV_API_BASE_URL + uri);
-    if (!response.success) {
-      throw "call to '/getMainPoolsGaugeRewards' didn't succeed";
+    // unfort these two endpoints have different keys for some pools
+    let searchKey = pool;
+    if (pool === 'y') {
+      searchKey = 'iearn';
+    } else if (pool === 'susd') {
+      searchKey = 'susdv2';
+    } else if (pool === 'ren2') {
+      searchKey = 'ren';
     }
-    return response.data.mainPoolsGaugeRewards;
-  } catch (error) {
-    return {};
-  }
-};
 
-const getPoolAPR = (pool, subgraph, gauge, crvPrice) => {
-  const crvPriceBN = BigNumber(crvPrice);
-  const decimals = BigNumber(1e18);
-  const workingSupply = BigNumber(gauge.gauge_data.working_supply).div(
-    decimals
-  );
-  const inflationRate = BigNumber(gauge.gauge_data.inflation_rate).div(
-    decimals
-  );
-  const relativeWeight = BigNumber(
-    gauge.gauge_controller.gauge_relative_weight
-  ).div(decimals);
+    const apyCrv = dataCrvApy[searchKey]?.crvApy;
 
-  const virtualPrice = BigNumber(subgraph.virtualPrice).div(decimals);
-  const nominator = pool.coins
-    .map((coin) => BigNumber(coin.poolBalance).times(coin.usdPrice))
-    .reduce((a, b) => a.plus(b));
-  const denominator = pool.coins
-    .map((coin) => BigNumber(coin.poolBalance))
-    .reduce((a, b) => a.plus(b));
-  const assetPrice = nominator.div(denominator);
-  try {
-    const poolAPR = crvPriceBN
-      .times(inflationRate)
-      .times(relativeWeight)
-      .times(12614400)
-      .div(workingSupply)
-      .div(assetPrice)
-      .div(virtualPrice)
-      .times(100);
-    return poolAPR.toNumber();
-  } catch (error) {
-    return 0;
-  }
-};
-
-const getPriceCrv = (pools) => {
-  //parse through pool coins and return price of crv dao token
-  for (const coin of pools.map((pool) => pool.coins).flat()) {
-    if (coin.address === '0xD533a949740bb3306d119CC777fa900bA034cd52') {
-      return coin.usdPrice;
+    // need to sum up the potential rewards apys
+    const apyRewardsArray = dataRewardApy[searchKey]?.additionalRewards;
+    let apyRewardsSum = 0;
+    if (apyRewardsArray?.length > 0) {
+      apyRewardsSum = apyRewardsArray
+        .map((el) => el.apy)
+        .reduce((accumulator, curr) => accumulator + curr);
     }
+    const apySum =
+      apyRewardsSum === undefined
+        ? apyBase + apyCrv
+        : apyBase + apyCrv + apyRewardsSum;
+
+    // 6) append output to object
+    poolStats.push({
+      pool,
+      virtual_price: poolData.virtual_price,
+      totalSupply: poolData.supply,
+      tvl,
+      apyBase,
+      apyCrv,
+      apyRewardsArray,
+      apyRewardsSum,
+      apy: apySum,
+    });
   }
+
+  // note(temporary only, investiage 4pool and 2pool apy)
+  return poolStats.filter((el) => el.pool !== '4pool' && el.pool !== '2pool');
+};
+
+const getDataEVM = async (chainString) => {
+  let tvl = await utils.getData(
+    `https://api.curve.fi/api/getTVL${
+      chainString.charAt(0).toUpperCase() + chainString.slice(1)
+    }`
+  );
+  tvl = tvl.data.pools;
+
+  const poolNames = Object.keys(tvl);
+  let dataTvl = [];
+  for (const p of poolNames) {
+    const obj = tvl[p];
+    obj['pool'] = p;
+    dataTvl.push(obj);
+  }
+
+  if (chainString === 'fantom') {
+    dataApy = await utils.getData(
+      `https://stats.curve.fi/raw-stats-ftm/apys.json`
+    );
+  } else {
+    dataApy = await utils.getData(
+      `https://stats.curve.fi/raw-stats-${chainString}/apys.json`
+    );
+  }
+  for (const el of dataTvl) {
+    el.apy = dataApy.apy.day[el.pool];
+  }
+
+  dataTvl = dataTvl.filter((el) =>
+    Object.keys(pools.tokenMapping[chainString]).includes(el.pool)
+  );
+
+  return dataTvl;
+};
+
+const buildPool = (entry, chainString) => {
+  const token = pools.tokenMapping[chainString][entry.pool];
+  const id = pools.pools[chainString].find(
+    (el) => el.name === entry.pool
+  )?.swap;
+  let apy = entry.apy < 0 ? 0 : entry.apy;
+  apy = chainString === 'ethereum' ? apy : apy * 100;
+
+  const newObj = {
+    // ugly, but some swap pools have same address on diff networks
+    pool: `${id}-${chainString}`,
+    chain: utils.formatChain(chainString),
+    project: 'curve',
+    symbol: utils.formatSymbol(token),
+    tvlUsd: entry.tvl,
+    apy: apy,
+  };
+  return newObj;
+};
+
+const topLvl = async (chainString) => {
+  if (chainString === 'ethereum') {
+    // pull data
+    poolStats = await getDataEth();
+  } else {
+    // pull data
+    poolStats = await getDataEVM(chainString);
+  }
+  // build pool objects
+  let data = poolStats.map((el) => {
+    if (pools.tokenMapping[chainString][el.pool] === undefined) {
+      return null;
+    }
+    return buildPool(el, chainString);
+  });
+
+  data = data.filter((el) => el !== null);
+  return data;
 };
 
 const main = async () => {
-  // request promises
-  const gaugePromise = getGaugesByChain();
-  const extraRewardPromise = getMainPoolGaugeRewards();
-  const blockchainToPoolSubgraphPromise = Object.fromEntries(
-    BLOCKCHAINIDS.map((blockchainId) => [
-      blockchainId,
-      getSubGraphData(blockchainId),
-    ])
-  );
-  const blockchainToPoolPromise = Object.fromEntries(
-    BLOCKCHAINIDS.filter(
-      (blockchainId) => blockchainId !== 'ethereum'
-    ).map((blockchainId) => [blockchainId, getPools(blockchainId)])
-  );
+  let data = await Promise.all([
+    topLvl('ethereum'),
+    topLvl('arbitrum'),
+    topLvl('avalanche'),
+    topLvl('fantom'),
+    topLvl('harmony'),
+    topLvl('optimism'),
+    topLvl('polygon'),
+  ]);
 
-  // we need the ethereum data first for the crv prive and await extra query to CG
-  const ethereumPools = await getPools('ethereum');
-  const priceCrv = getPriceCrv(Object.values(ethereumPools));
-
-  // create feeder closure to fill defillamaPooldata asynchroniously
-  const defillamaPooldata = [];
-  const feedLlama = (poolData, blockchainId) => {
-    const [
-      addressToPool,
-      addressToPoolSubgraph,
-      addressToGauge,
-      gaugeAddressToExtraRewards,
-    ] = poolData;
-    for (const [address, pool] of Object.entries(addressToPool)) {
-      const subgraph = addressToPoolSubgraph[address];
-      const gauge = addressToGauge[blockchainId][pool.gaugeAddress];
-      // one gauge can have multiple (different) extra rewards
-      const extraRewards = gaugeAddressToExtraRewards[pool.gaugeAddress];
-
-      const apyBase = subgraph ? parseFloat(subgraph.latestDailyApy) : 0
-      const aprCrv =
-        gauge && subgraph ? getPoolAPR(pool, subgraph, gauge, priceCrv) : 0;
-      const aprExtra = extraRewards
-        ? extraRewards.map((reward) => reward.apy).reduce((a, b) => a + b)
-        : 0;
-
-      // tokens are listed using their contract addresses
-      // https://github.com/DefiLlama/yield-server#adaptors
-      const underlyingTokens = pool.coins.map((coin) => coin.address);
-      const rewardTokens = extraRewards ? extraRewards.map(reward => reward.tokenAddress) : []
-      if (aprCrv) {
-        rewardTokens.push('0xD533a949740bb3306d119CC777fa900bA034cd52') // CRV
-      }
-
-      const tvlUsd = pool.usdTotal
-      if (!tvlUsd) {
-        continue
-      }
-
-      defillamaPooldata.push({
-        pool: address + '-' + blockchainId,
-        chain: utils.formatChain(blockchainId),
-        project: 'curve',
-        symbol: pool.coins.map((coin) => coin.symbol).join('-'),
-        tvlUsd,
-        apyBase,
-        apyReward: aprCrv + aprExtra,
-        apy: apyBase + aprCrv + aprExtra,
-        rewardTokens,
-        underlyingTokens,
-      });
-    }
-  };
-
-  // group Promises by blockchain and feed the llama array
-  const responses = [];
-  for (const [blockchainId, poolSubgraphPromise] of Object.entries(
-    blockchainToPoolSubgraphPromise
-  )) {
-    // we already have ethererum for the crv price
-    if (blockchainId !== 'ethereum') {
-      responses.push(
-        Promise.all([
-          blockchainToPoolPromise[blockchainId],
-          poolSubgraphPromise,
-          gaugePromise,
-          extraRewardPromise,
-        ]).then((poolData) => feedLlama(poolData, blockchainId))
-      );
-    } else {
-      responses.push(
-        Promise.all([
-          poolSubgraphPromise,
-          gaugePromise,
-          extraRewardPromise,
-        ]).then((APYandGaugeData) =>
-          feedLlama([ethereumPools, ...APYandGaugeData], blockchainId)
-        )
-      );
-    }
-  }
-
-  // wait for all Group promises to resolve
-  try {
-    await Promise.all(responses);
-  } catch (error) {
-    console.error(error);
-  }
-
-  return defillamaPooldata;
+  return data.flat();
 };
 
 module.exports = {
   timetravel: false,
   apy: main,
-  // legacy needed by convex-finance adaptor
-  curvePoolStats: legacy.curvePoolStats,
-  tokenMapping: legacy.tokenMapping['ethereum'],
+  curvePoolStats: getDataEth,
+  tokenMapping: pools.tokenMapping['ethereum'],
 };
